create program search componenet
