+++ src/pages/program-search.tsx
@@ -1,0 +1,8 @@
+import ProgramFilter from "@component/components/ProgramSearch/ProgramFilter";
+import { NextPage } from "next";
+
+const ProgramSearch: NextPage = () => {
+  return <ProgramFilter />;
+};
+
+export default ProgramSearch;
